# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AprovaFácil API",
    "description": "API para sistema de estudos e preparação para concursos",
    "version": "1.1.0",
    "contact": {
      "name": "AprovaFacil Team",
      "email": "contato@aprovafacil.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Servidor de desenvolvimento"
    },
    {
      "url": "https://api.aprovafacil.com",
      "description": "Servidor de produção"
    }
  ],
  "paths": {
...
    "/guru/v1/activities/flashcards": {
      "get": {
        "operationId": "getGuruActivitiesFlashcardsV1",
        "summary": "Guru v1 - Atividades de flashcards",
        "description": "Lista as atividades recentes relacionadas a flashcards.",
        "tags": [
          "Guru"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Quantidade máxima de atividades",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 10
            }
          },
          {
            "$ref": "#/components/parameters/CorrelationId"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/GuruActivity"
                      }
                    }
                  }
                }
              }
            },
            "headers": {
              "x-correlation-id": {
                "description": "ID de correlação",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit excedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "CorrelationId": {
        "name": "x-correlation-id",
        "in": "header",
        "required": false,
        "description": "ID de correlação para rastreamento ponta a ponta",
        "schema": {
          "type": "string"
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Mensagem de erro"
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "ERROR_CODE"
              },
              "message": {
                "type": "string",
                "example": "Mensagem de erro"
              }
            }
          }
        }
      },
...
      "GuruActivity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "example": "simulado"
          },
          "titulo": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          },
          "time": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "score": {
            "type": "number",
            "nullable": true
          },
          "improvement": {
            "type": "number",
            "nullable": true
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}